+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.10.6-dev
+## 0.10.6
 
 * Enable Dart 2 Preview in analyzer options (including non-Flutter packages).